Moved authorization code into its own blueprint.
